corpus_explanation/requirements.txt
#pip install torch===1.3.1 torchvision===0.4.2 -f https://download.pytorch.org/whl/torch_stable.html
modelsummary==1.1.7
torchsummary
torchtext==0.3.1
spacy==2.1.9
nltk==3.4.5
rake-nltk==1.0.4
bs4==0.0.1
tqdm==4.40.2
sklearn==0.22.1
# python -m spacy download en
